Refactor validator querying and add utility functions for validator selection
