+++ folding_api/chain.py
@@ -8,6 +8,8 @@ class SubtensorService:
         # Initialize subtensor with config
         self.subtensor = bt.subtensor(config=self.config) if config else bt.subtensor()
         self.metagraph = self.subtensor.metagraph(self.config.netuid if config else 25)
+        self.wallet = bt.wallet(config=self.config) if config else bt.wallet()
+        self.dendrite = bt.dendrite(wallet=self.wallet)
 
     def resync_metagraph(self):
         self.metagraph.sync(subtensor=self.subtensor)
